Fixed playRandomTrack()
Tested method with an empty tracklist, and bad things happened.
So, added an if statement to test size of arraylist.
Also figured out that random.nextInt starts at 0, not at 1, so I didn't need to subtract 1 from the random integer range to get valid index numbers.
With that, I think 4.43 is completed.