Yikes
The player's too smart!
I'm trying to figure out how to get the player to wait until a song is finished before playing the next one.
The "wait until done and return" thing is in the playSample method, but not in the startPlaying method...
Which on a user level is nice. It means I can do things while the music is playing, but it makes it a pain for coding.
Conundrums...