Attempting to tackle 4.45
Doing this problem using playSample rather than startPlaying() until I figure out how to get the program to wait for a song to be over before starting the next one.
Trying to go the route of copying the original track list, playing a random song, removing it from the new list, and repeat.
I'm running into the problem that the method only plays one song and then throws an error. I think this happens because the index is removed at the end and when it goes to loop around it has no reference for where it left off.