Figured it out for 4.45
I was stuck on using a for each loop.
Changed it to a while loop that simply tests the size of the new shufflePlay arrayList. Which worked wonders :)
Now, however, I've run into the problem that since playSample() doesn't return until after the song finishes and even then it returns to a loop, the stopPlaying method doesn't work.